+++ src/frontend/components/components/ComponentViewerSkeleton.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Skeleton } from '@/components/ui/skeleton'
-import { Card } from './ui/card'
+import { Card } from '@/components/ui/card'
 
 export default function ComponentViewerSkeleton({
   message